+++ slack-gamebot/api/presenters/team_presenter.rb
@@ -14,6 +14,7 @@ module Api
       property :gifs, type: Boolean, desc: 'Team loves animated GIFs.'
       property :aliases, type: Array, desc: 'Game aliases.'
       property :elo, type: Integer, desc: 'Base elo.'
+      property :unbalanced, type: Boolean, desc: 'Permits unbalanced challenges.'
       property :created_at, type: DateTime, desc: 'Date/time when the team was created.'
       property :updated_at, type: DateTime, desc: 'Date/time when the team was accepted, declined or canceled.'