+++ slack-gamebot/models/team.rb
@@ -4,6 +4,7 @@ class Team
   field :aliases, type: Array, default: []
   field :nudge_at, type: DateTime
   field :elo, type: Integer, default: 0
+  field :unbalanced, type: Boolean, default: false
 
   field :stripe_customer_id, type: String
   field :premium, type: Boolean, default: false